+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated fox sports extraction to use new `Runner` and `Scraper` pattern (#80)
 - Additional validation in fox sports extraction (#80)
 - Updated baseball reference mlb extraction to use new `Runner` and `Scraper` pattern
+- Updated basketball reference nba extraction to use new `Runner` and `Scraper` pattern
+- Code organization improvements
 
 ## [0.10.0] - 2025-05-29
 ### Changed